chore: 장바구니의 특정 메뉴를 삭제할 때 정말 삭제할 것인지 confirm 한다.

+++ src/CartContainer.jsx
@@ -47,7 +47,9 @@ export default function CartContainer() {
   };
 
   const handleClickRemoveCartItem = (menuId) => {
-    dispatch(requestRemoveCartItem(menuId));
+    if (window.confirm('메뉴를 삭제하시겠습니까?')) {
+      dispatch(requestRemoveCartItem(menuId));
+    }
   };
 
   const handleClickOrder = () => {
